+++ file_config/handlers/json.py
@@ -10,12 +10,13 @@ class JSONHandler(BaseHandler):
 
     name = "json"
     packages = ("ujson", "json")
+    options = {"indent": 0}
 
-    def on_json_dumps(self, json, dictionary):
+    def on_json_dumps(self, json, dictionary, **kwargs):
         """ The :mod:`json` dumps method.
         """
 
-        return json.dumps(dictionary)
+        return json.dumps(dictionary, indent=kwargs.get("indent", 0))
 
     def on_json_loads(self, json, content):
         """ The :mod:`json` loads method.
@@ -23,11 +24,11 @@ class JSONHandler(BaseHandler):
 
         return json.loads(content)
 
-    def on_ujson_dumps(self, ujson, dictionary):
+    def on_ujson_dumps(self, ujson, dictionary, **kwargs):
         """ The `ujson <https://pypi.org/project/ujson/>`_ dumps method.
         """
 
-        return ujson.dumps(dictionary)
+        return ujson.dumps(dictionary, indent=kwargs.get("indent", 0))
 
     def on_ujson_loads(self, ujson, content):
         """ The `ujson <https://pypi.org/project/ujson/>`_ loads method.